Rename domain name anonymous -> mentee
