+++ src/main/java/site/hirecruit/hr/domain/mentee/service/MenteeService.kt
@@ -1,0 +1,4 @@
+package site.hirecruit.hr.domain.mentee.service
+
+interface MenteeService {
+}
