+++ src/main/java/site/hirecruit/hr/domain/mentoring/service/MentoringService.kt
@@ -1,0 +1,5 @@
+package site.hirecruit.hr.domain.consultation.service
+
+interface ConsultationService {
+    fun RequestConsultationWithMentor()
+}
